+++ CHANGELOG.md
@@ -3,8 +3,8 @@
 ## 0.31.0 - TBD
 
 ### Enhancements
-- Added new venues, datasets, and publishers for ICE Futures US and for ICE Futures
-  Europe (Financial products)
+- Added new venues, datasets, and publishers for ICE Futures US, ICE Futures Europe
+  (Financial products), Eurex, and European Energy Exchange (EEX)
 
 ## 0.30.0 - 2025-02-11